Bug fix: bowls not disappearing properly
First the bowls disappeared from one side of the shelf en masse when hit
with the ball. By determining whether the collider is a mesh or box type
I fixed that.

However the mesh colliders wouldn't disappear when the ball hit it, so I
had to traverse the transforms in the hierarchy to find the Mesh
Collider and disable it.