Added a bit of sound
Due to the discrepancies tracking the prefabs, the sound effects didn't
apply to every instance of bowl and square plate. They broke the thin
shelves as well - i.e. AABB wasn't found or failed - and caused the
objects to lose their mesh renderers.

This game is so broken right now.